update: 数据类型修改和bug修复
- 更新数据库映射字段为Long类型，以提高数值范围和一致性。- 实现MybatisSql日志拦截器，优化SQL查询并记录执行信息。
- 修复异常处理，优化日志记录策略，提升系统稳定性和可维护性。

+++ yunju-club-subject/yunju-club-application/yunju-club-application-controller/src/main/java/com/zsyj/subject/application/dto/SubjectMappingDTO.java
@@ -18,11 +18,11 @@ public class SubjectMappingDTO implements Serializable {
     /**
      * 分类id
      */
-    private Integer categoryId;
+    private Long categoryId;
     /**
      * 标签id
      */
-    private Integer labelId;
+    private Long labelId;
 
 
 }
