+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/entity/SubjectMapping.java
@@ -24,11 +24,11 @@ public class SubjectMapping implements Serializable {
     /**
      * 分类id
      */
-    private Integer categoryId;
+    private Long categoryId;
     /**
      * 标签id
      */
-    private Integer labelId;
+    private Long labelId;
     /**
      * 创建人
      */
