+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/mapper/SubjectLabelDao.java
@@ -77,8 +77,8 @@ public interface SubjectLabelDao {
      * @param labelIds labelId
      * @return list List<SubjectLabel>
      */
-    List<SubjectLabel> querySubjectLabelById(List<Integer> labelIds);
+    List<SubjectLabel> querySubjectLabelById(List<Long> labelIds);
 
-    List<SubjectLabel> batchQueryById(List<Integer> labelIdList);
+    List<SubjectLabel> batchQueryById(List<Long> labelIdList);
 }
